빌드 경고를 0 으로 — 대부분은 리팩터가 남긴 흔적이었다
클린 리빌드에서 Debug 150 건, Release 6 건. 125 건이 -Wdocumentation 이었는데
전부 같은 모양이다: 선언이 옮겨지거나 지워졌는데 주석만 제자리에 남아 다음
선언 위에 얹힌 것. 그러니까 "문서가 엉뚱한 것을 설명하고 있다" 는 신고였다.
GpuMaterialGroup 설명은 GpuMaterialElementKey 를, GpuBatchInfo 설명은
GpuBatchSortMode 를, materialGroupFor 설명은 retireUnusedMaterialElements 를
설명하는 자리에 있었다. 전부 자기 선언 옆으로 돌려놨다.

경고가 실제 결함을 짚은 것 둘:

- kDefaultKeyLightColor 가 안 쓰이고 있었다. 주광 기본값이 FrameLightState 로
  옮겨 갔는데 .cpp 의 복사본이 남은 것인데, **방향 쪽은 더 나빴다** —
  buildLightViewProj 만 여전히 .cpp constexpr 을 봐서, 패킷이 다른 빛 방향을
  실어 주면 셰이딩과 그림자 행렬이 서로 다른 빛을 보게 돼 있었다.
  이제 둘 다 _frameLight 를 읽는다.
- ReloadFileManager 의 pActionStr 가드가 `!SW_SHIPPING` 이었다. Release 는
  Shipping 이 아니면서 Trace 를 컴파일하지 않으므로, 문자열만 만들고 쓰는 데가
  없었다. 가드를 SW_LOG_LEVEL_COMPILED(3) 으로 바꿔 로그와 생사를 같이 하게 했다.

나머지:
- offsetMap.find( 0 ) 이 uint32 키 맵을 부호 있는 리터럴로 찾아 이종 키
  오버로드로 새고 있었다 (0u).
- Vulkan 밴드 판정의 `binding >= kBShift` 는 kBShift 가 0 이라 항상 참이었다.
  부호 없는 뺄셈 한 번으로 상·하한을 같이 보는 inBand 로.
- DXGI_FORMAT 은 값이 110 개가 넘는 플랫폼 enum 이라 -Wswitch-enum 을 만족시킬
  방법이 없다(새 포맷이 생기면 Unknown 이 맞는 답이다). 그 함수만 국소 pragma 로
  끈다 — 반대 방향인 toDxgiFormat 은 우리 enum 이므로 경고를 그대로 받는다.
- GL 의 두 포맷 스위치에 BC 계열을 명시했다. "압축은 glCompressedTexImage 경로로
  간다" 를 default 에 묻어 두지 않고 적는다.
- 리플렉션 코드젠이 float32 멤버에 double 리터럴을 냈다. 접미사 f 를 붙여
  생성 파일마다 나던 경고를 한 자리에서 없앴다.
- TestLog 의 ThreadSafeLogCapture 는 Debug 전용 테스트에서만 쓰는데 정의가 가드
  밖에 있었다. 쓰는 쪽과 같은 가드 안으로.

검증: Ninja-Debug / Ninja-Release 둘 다 클린 리빌드 경고 0,
ctest -L nogpu 5/5, RenderPassTest 33/33, RHITest 13/13, Core_Log 13/13,
벤치 drawn=6811 (기준 6811~6832).

+++ Test/EngineTest/TestRenderPass.cpp
@@ -1508,7 +1508,7 @@ SW_TEST_CASE( RenderPassTest, MainPassCullsWithCameraFrustumNotLight )
             for ( uint32 sideIndex = 0; sideIndex < 2 && bOk; ++sideIndex )
             {
                 sw::string      name = sw::string( "FarCube" ) + sw::to_string( sideIndex );
-                sw::GameObject* pObj = scene.getObjectManager()->createGameObject( sw::hashed_string( name.c_str(), name.size() ) );
+                sw::GameObject* pObj = scene.getObjectManager()->createGameObject( sw::hashed_string( name.c_str(), static_cast<uint32>( name.size() ) ) );
                 bOk                  = pObj != nullptr;
                 if ( bOk == false )
                     break;
@@ -1638,7 +1638,7 @@ SW_TEST_CASE( RenderPassTest, TransparentOrderMatchesAcrossBackends )
             for ( uint32 cubeIndex = 0; cubeIndex < kCubeCount && bOk; ++cubeIndex )
             {
                 sw::string      name = sw::string( "Glass" ) + sw::to_string( cubeIndex );
-                sw::GameObject* pObj = scene.getObjectManager()->createGameObject( sw::hashed_string( name.c_str(), name.size() ) );
+                sw::GameObject* pObj = scene.getObjectManager()->createGameObject( sw::hashed_string( name.c_str(), static_cast<uint32>( name.size() ) ) );
                 bOk                  = pObj != nullptr;
                 if ( bOk == false )
                     break;
@@ -1814,7 +1814,7 @@ SW_TEST_CASE( RenderPassTest, GpuGeneratedCommandsDrawOnlyVisibleInstances )
             for ( uint32 objectIndex = 0; objectIndex < kObjectCount && bOk; ++objectIndex )
             {
                 sw::string      name = sw::string( "CullCube" ) + sw::to_string( objectIndex );
-                sw::GameObject* pObj = scene.getObjectManager()->createGameObject( sw::hashed_string( name.c_str(), name.size() ) );
+                sw::GameObject* pObj = scene.getObjectManager()->createGameObject( sw::hashed_string( name.c_str(), static_cast<uint32>( name.size() ) ) );
                 bOk                  = pObj != nullptr;
                 if ( bOk )
                 {
@@ -2231,7 +2231,7 @@ SW_TEST_CASE( RenderPassTest, FrameRendererParityAllBackends )
                 break;
 
             sw::string      objectName = sw::string( "Cube" ) + sw::to_string( meshIndex );
-            sw::GameObject* go         = scene.getObjectManager()->createGameObject( sw::hashed_string( objectName.c_str(), objectName.size() ) );
+            sw::GameObject* go         = scene.getObjectManager()->createGameObject( sw::hashed_string( objectName.c_str(), static_cast<uint32>( objectName.size() ) ) );
             bOk                        = go != nullptr;
             if ( bOk )
             {
